dracut/30ignition/ignition-setup.sh: migrate to Ignition v3
In Ignition v3 we don't have a "default" config anymore and there can
now be multiple "base" configurations but they have to be in a
subfolder.

Skip copying the "default" config, which should be repaced by a new OEM
systemd service that checks whether Ignition found a config or not.
Use the new "base.d" subfolder as target for the OEM "base" config.