Port to Ignition v3
The dracut/30ignition files had to be updated to match the folder
dracut/30ignition found in the ignition repo but with some adaptions
because Flatcar allows to overwrite the root filesystem and has a
different initramfs service setup.
For Ignition v3 we had to provide an implementation for is-live-image
and ignition-kargs-helper. The kargs helper does not handle all cases
due to the complexity of evaluation grub.cfg.