ignition-kargs-helper: Ensure that grub.cfg exists before reading
The kola kargs test fails for the generic image when it's not mangled
(which is the kola term for setting the console option by creating the
grub.cfg file) because the kargs helper didn't handle the case where
the file doesn't exist which is normal for the generic image.
Create the grub.cfg file if it doesn't exist before trying to read from
it.