Manage systemd-sysext images for A/B booting
Extensions to the generic /usr partition were in the done through a
base Ignition config and extra files on the OEM partition which weren't
updated. With systemd-sysext images we have a way to extend the /usr
partition through an overlay and have single file updates.

Define a mechanism to discover and activate OEM sysext images and
other optional Flatcar extensions that are coupled to the OS version
and get updated by update-engine in an A/B scheme.
The OEM sysext images are stored on the OEM partition as usual because
they may be required in airgapped environments and users may want to
reformat the root partition with Ignition in which case storing it on
the rootfs or downloading it would not work well. If we have size
limitations the inactive sysext gets moved to the rootfs but the active
sysext is still placed on the OEM partition unless this also doesn't
fit and users of those platforms that want to reformat the root
partition would need to ensure that downloading works for them or they
provide the file out of band with Ignition.
Other optional Flatcar extensions are only stored on the rootfs and
are always downloaded unless a special baked image is used or the file
is provided through Ignition.
This change introduces a new phase for the rootfs preparation that runs
after Ignition. Both the phases before and after Ignition will run on
every boot. The existing rootfs preparation had an initrd.target.wants
link but that didn't work and is now removed.
The RequiresMountsFor=/sysroot/usr/share/oem directive didn't have any
effect because there was no mount unit defined. The directive
After=sysroot-usr.mount is already implied by
RequiresMountsFor=/sysroot/usr/.