initrd-setup-root: Run always to ensure rootfs is populated
In https://github.com/flatcar/Flatcar/issues/944 we noticed that dbus
failed because of missing files in the rootfs when Ignition isn't
running in a PXE environment. This worked before by chance but the
underlying problem is that initrd-setup-root is required but wasn't
running because the "wants" symlinks in the module setup weren't
taking effect.

This is a backport of the relevant changes in
https://github.com/flatcar/bootengine/pull/50 to always run
initrd-setup-root instead of only when Ignition runs by pulling it in
from the two possible targets: either ignition-complete or
ignition-subsequent which runs in the other case.
The RequiresMountsFor=/sysroot/usr/share/oem directive didn't have any
effect because there was no mount unit defined. The directive
After=sysroot-usr.mount is already implied by
RequiresMountsFor=/sysroot/usr/.