Fix non-working norecovery option
The check for the norecovery option used the same condition but had
different actions in the if branch. Also, for the systemd generator
the unit name was wrong because it was just the unit name of the bind
mount.
Align the check for the norecovery option and also use the right unit
name.