initrd-setup-root: Clean duplicate overlayfs entries
With machines having an old full /etc they would not benefit from
future updates of files in /usr/share/flatcar/etc.
Clean up duplicate entries to make sure that the lowerdir gets used.
While this could be done in a one-time migration it is actually not bad
to do this also against unwanted upcopies. The walker helper function
introduced should also be used for the selective OS reset feature.