initrd-setup-root: Make script more resilient and handle deleted passwd
A rerun from the initrd due to manual invocation could easily result in
a failure of the mount if it is already mounted or in working on merged
overlay files instead of only the upperdir. Another similar failure
case is when this wrong usage or even a deletion of /etc/passwd on the
booted system results in a wiped out entry for /etc/passwd (overlayfs
uses a special character device in the upperdir) which flatcar-tmpfiles
can't recreate.
Support these edge cases as they seemed to have worked before and it
also makes development/debugging easier.