Prevent the OEM mount point from showing up before Ignition mounts
Ignition may run a manual mount command for the OEM partition if the
user specifies it. This fails if the OEM mount unit is triggered early,
because it can't be mounted twice, thus we need to prevent that from
happening by unmounting the OEM partition (but not the tmpfs OEM mount
for PXE-provided OEM contents). To spare users from even dealing with
the OEM partition specification in Ignition we can ensure it is mounted
before Ignition writes files.