dracut: Move OEM partition mountpoint from /usr/share/oem to /oem
This is to move the OEM partition outside the /usr partition to enable
OEM sysext. systemd unmounts everything under /usr when creating the
overlay, which means that the OEM partition also gets unmounted and
this makes sysext images stored inside it unavailable. This results in
no sysext image to be deployed.

The only /usr/share/oem path that was left as is is in initrds.