setup-root: Detect initial sysext MVP
We have implemented the building of OEM sysext images but the update
path is still in progress. Instead of blocking the migration for new
deployments, we will lose the version coupling by using "initial" as
version in the file name and SYSEXT_LEVEL instead of VERSION_ID for the
extension release file.
Detect the initial sysext MVP path to activate the OEM sysext.