setup-root: Read old OEM files from the migration flag file
The list of old files per OEM needs to be maintained somewhere. Instead
of doing this in the initrd script, we can use the update-engine
post-inst action which already is a place for migration logic.
Expect that the post-inst action writes all old OEM files to clean up to
the migration flag file that triggers the deletion.