setup-root: Prevent a bootloop on service failure
Depending on the timing, the failure of the setup-root service may cause
a boot loop when the emergency target starts but the service continues
to restart. This was seen with the coreos.update.badusr kola test.
As with other important services in the initrd, make sure we start the
emergency target in the isolate job mode which stops all other services.