setup-root: Download sysext images if required
The sysext images required for the OEM or optional extensions should be
downloaded if not found. For OEM sysext images they can be missing if
the user deleted them or if the additional update payload didn't wasn't
used (e.g., a custom Nebraska server or flatcar-update without the right
OEM payload). For optional extensions it is expected that they get
downloaded on-demand on first boot (like with the OEM sysext we also
need a fallback)
Add download logic for OEM sysext images as fallback method. Also use
the same download logic for optional extensions where they get
downloaded on first-boot (or on reboot as fallback when the update tool
didn't provide the right payload).