initrd-setup-root: Add workaround for missing DNS resolve retry
When the first DNS resolving fails, curl goes into the retry loop but
reuses the failed result, thus making the retry pointless. This can
happen when it races with the DNS setup being complete or when a
temporary failure occurs (The script waits for resolved to start but
that doesn't fully guarantee that the first resolving directly works).
Add a pre-waiter that retries DNS resolving.