initrd-setup-root: Migrate cloud-init enablement symlink
The old OEM setup was creating and enabling a cloud-init service unit
from the default.ign file on the rootfs. When deleting this, we need
to ensure that cloud-init is still enabled by enabling the inbuilt
service unit from the base image.
Special-case the removal of the cloud-init enablement symlink to create
it again with the new target.