initrd-setup-root: Improve Flatcar extension config format
A simple list of extensions to enable is not enough in case we want to
move parts of the base image to an extension because then we would have
to enable it by default. This requires an option for the user to
explicity disable it.
Add support for reading the OS default extensions from /usr and allow
that the user disables extensions through writing "-NAME" in the
extension config file in /etc.