initrd-setup-root-after-ignition: Use update key for initrd sysext DL
The OEM and "official" Flatcar systemd-sysext images have a download
mechanism in the initrd. For OEM extensions this is only a fallback in
case the user deleted the files because now we have all things in place
for the update to ensure that OEM payloads are there. The enabling of
"official" Flatcar extensions on provisioning still would primarily go
through this mechanism and therefore, it should not use the image
signing key but the update key to have a single way of verifying the
extensions.
Switch the extension download to use the update server and the update
key for verification of the payloads.