ignition: Drop sysroot-boot.service in favour of boot.mount
sysroot-boot.service used to be a simple mount unit way back before
8bf46fe78ec59bcd5148ce9ab8ec5fb805600151. This was changed due to umount
sometimes being called after an unmount had already happened, causing it
and its associated unit to fail. This was harmless, but it did cause
Kola test failures.

Although the precise cause of that issue was never found, I think that
between moving the mount point outside of /sysroot, not forcing it to
unmount before switching root (that will happen anyway), and the
intervening 8 years of systemd changes, there's a good chance we won't
see it again.

Signed-off-by: James Le Cuirot <jlecuirot@microsoft.com>