Remove unused generator code
The minimal initrd now does the verity setup and /sysusr mounting and we
can delete parts of the generator code.
Delete the unused code but still keep support for "usr=" instead of
"mount.usr=" cmdline entries. We could decide to remove this later to
get rid of a few more things in the generators. Still, /sysroot(/usr)
setup will be needed. For "mount.usr=" the upstream systemd generator
handles this but there's still a bit of code to customize the mount args
for it. The verity generator gets deleted fully and if we later want to
use it again, we can use the upstream one with some cmdline
preprocessing.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>