minimal-init: Try to load modules for all devices again after a timeout
Normally it works to do one full module load and then the drive is
detected but in case this takes longer and mdev doesn't do the job of
module loading we can fall back to a full load again to try to progress.
So far we don't know if this helps (it won't for completly missing
kernel modules of course) but it's good to have this done automatically
instead of having to ask users if it helps for them when they hit an
issue.

Signed-off-by: Kai Lueke <kailuke@microsoft.com>